+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Removed
 
-- Remove owner_id from group summary ([#3337](https://github.com/open-ic/open-chat/pull/3337))
+- Remove owner_id from group summary ([#3340](https://github.com/open-ic/open-chat/pull/3340))
 
 ### Changed